Abstracted out standalone functions from Covid19 for better testing.
